adicionando hábito

+++ src/main/resources/db/migration/V2__create-table-habit.sql
@@ -4,7 +4,7 @@ CREATE TABLE habit (
     done BOOLEAN NOT NULL,
     start DATE NOT NULL,
     end DATE NOT NULL,
-    percentage_for_day FLOAT(3, 2),
+    percentage_for_day FLOAT(5, 2),
     client_id BIGINT     NOT NULL,
     PRIMARY KEY (id),
     CONSTRAINT fk_client FOREIGN KEY (client_id) REFERENCES client(id) ON DELETE CASCADE
